Fold S3 code into persistent module
