+++ src/lib.rs
@@ -1,3 +1,2 @@
 pub mod error;
 pub mod persistent;
-pub mod s3;
